Add files via upload
The Urban Power Consumption Monitoring Project in Eastern China (UPCMP) is a dataset that samples the electricity consumption of 239 customers in a city in eastern China. The dataset is released publicly with the consent of the relevant authorities after the raw load data has been max-min normalized and the actual ID information of customers has been removed. The data is sampled at a time granularity of 15 minutes/time and collected over the period of 16 January 2020 - 28 March 2020 and 31 July 2020 - 11 September 2020. In addition to customer electricity consumption information, the UPCMP also gives the daily maximum and minimum temperatures of the city during the sampling period.